+++ connect-for-commerce-scope-server/src/main/java/to/be/renamed/module/projectconfig/model/ProjectAppConfiguration.java
@@ -15,10 +15,11 @@ public class ProjectAppConfiguration implements Serializable {
     private final GeneralConfig generalConfig;
     private final BridgeConfig bridgeConfig;
     private final ReportConfig reportConfig;
+    private final JwtConfig jwtConfig;
 
     @SuppressWarnings("unused") // Required by Gson
     private ProjectAppConfiguration() {
-        this(null, null, null);
+        this(null, null, null, null);
     }
 
     /**
@@ -28,10 +29,11 @@ public class ProjectAppConfiguration implements Serializable {
      * @param bridgeConfig  Bridge configuration
      * @param reportConfig  Report configuration
      */
-    public ProjectAppConfiguration(GeneralConfig generalConfig, BridgeConfig bridgeConfig, ReportConfig reportConfig) {
+    public ProjectAppConfiguration(GeneralConfig generalConfig, BridgeConfig bridgeConfig, ReportConfig reportConfig, JwtConfig jwtConfig) {
         this.generalConfig = Objects.requireNonNullElse(generalConfig, new GeneralConfig());
         this.bridgeConfig = Objects.requireNonNullElse(bridgeConfig, new BridgeConfig());
         this.reportConfig = Objects.requireNonNullElse(reportConfig, new ReportConfig());
+        this.jwtConfig = Objects.requireNonNullElse(jwtConfig, new JwtConfig());
     }
 
     public GeneralConfig getGeneralConfig() {
@@ -46,9 +48,14 @@ public class ProjectAppConfiguration implements Serializable {
         return reportConfig;
     }
 
+    public JwtConfig getJwtConfig() {
+        return jwtConfig;
+    }
+
     @Override
     public String toString() {
-        return format("{generalConfig: [%s], bridgeConfig: [%s], reportConfig: [%s]}", generalConfig, bridgeConfig, reportConfig);
+        return format("{generalConfig: [%s], bridgeConfig: [%s], reportConfig: [%s], jwtConfig: [%s]}", generalConfig, bridgeConfig, reportConfig,
+                      jwtConfig);
     }
 
     @Override
@@ -64,11 +71,12 @@ public class ProjectAppConfiguration implements Serializable {
         ProjectAppConfiguration that = (ProjectAppConfiguration) o;
         return Objects.equals(getGeneralConfig(), that.getGeneralConfig()) &&
                Objects.equals(getBridgeConfig(), that.getBridgeConfig()) &&
-               Objects.equals(getReportConfig(), that.getReportConfig());
+               Objects.equals(getReportConfig(), that.getReportConfig()) &&
+               Objects.equals(getJwtConfig(), that.getJwtConfig());
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(getGeneralConfig(), getBridgeConfig(), getReportConfig());
+        return Objects.hash(getGeneralConfig(), getBridgeConfig(), getReportConfig(), getJwtConfig());
     }
 }